refactor(page-header): add utils file
