+++ packages/ibm-products-web-components/src/components/page-header/page-header.ts
@@ -14,71 +14,7 @@ import { prefix } from '../../globals/settings';
 import styles from './page-header.scss?lit';
 import { carbonElement as customElement } from '@carbon/web-components/es/globals/decorators/carbon-element.js';
 import { pageHeaderContext } from './context';
-
-/**
- * ----------
- * Utilities
- * ----------
- */
-
-const getHeaderOffset = (el: HTMLElement): number => {
-  const scrollableContainer = scrollableAncestor(el);
-  const scrollableContainerTop = scrollableContainer
-    ? (scrollableContainer as HTMLElement).getBoundingClientRect().top
-    : 0;
-  const offsetMeasuringTop = el ? el.getBoundingClientRect().top : 0;
-  const totalHeaderOffset =
-    offsetMeasuringTop !== 0 ? offsetMeasuringTop - scrollableContainerTop : 0;
-  return totalHeaderOffset;
-};
-
-const windowExists = typeof window !== `undefined`;
-
-/**
- * Determines if the given target is scrollable
- *
- * @param {HTMLElement} target
- * @returns {boolean}
- */
-const scrollable = (target: HTMLElement): boolean => {
-  const style = window.getComputedStyle(target);
-  return /(auto|scroll|hidden)/.test(style.overflow);
-};
-
-/**
- * Recursively looks for the scrollable ancestor
- */
-export const scrollableAncestorInner = (target: HTMLElement) => {
-  if (target.parentNode && target.parentNode !== document) {
-    if (scrollable(target.parentNode as HTMLElement)) {
-      return target.parentNode;
-    } else {
-      return scrollableAncestorInner(target.parentNode as HTMLElement);
-    }
-  } else {
-    return document.scrollingElement;
-  }
-};
-
-/**
- * Walks up the parent nodes to identify the first scrollable ancestor
- *
- * @param {HTMLElement} target
- * @returns {HTMLElement}
- */
-const scrollableAncestor = (target: HTMLElement) => {
-  if (!windowExists || !target) {
-    return null;
-  }
-
-  // based on https://stackoverflow.com/questions/35939886/find-first-scrollable-parent
-  const style = window.getComputedStyle(target);
-
-  if (!target || !style || style.position === 'fixed') {
-    return document.scrollingElement;
-  }
-  return scrollableAncestorInner(target);
-};
+import { getHeaderOffset } from './utils';
 
 export interface offsetValues {
   breadcrumbOffset?: number;
